Refactor suggested by Marcial
